refactor(theme): refactor SearchBar component - move search bar component to components/Header - remove unused code and watchers
